fixing warnings

+++ database.h
@@ -55,8 +55,8 @@ public:
 	bool drop(const char *document_id, bool commit);
 	char* prefixed(const char *term, const char *prefix);
 	bool index(const char *document, bool commit);
-	long long int get_slot(const char *name);
-	long long int hex2long(const char *input);
+	unsigned int get_slot(const char *name);
+	unsigned int hex2long(const char *input);
 	char* _date(const char *iso_date);
 	char* _time(const char *iso_date);
 	char* _sTZD(const char *iso_date);
